+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.3 - in progress
 
 - added tests for findProxiesHostPort
+- added values to queriesForSearchEngine
 
 ## 0.3.2 - 2021.08.14